+++ Source/Core/Credit.js
@@ -2,11 +2,13 @@ define([
     './defaultValue',
     './defined',
     './defineProperties',
+    './deprecationWarning',
     './DeveloperError'
 ], function(
     defaultValue,
     defined,
     defineProperties,
+    deprecationWarning,
     DeveloperError) {
     'use strict';
 
@@ -15,11 +17,11 @@ define([
 
     /**
      * A credit contains data pertaining to how to display attributions/credits for certain content on the screen.
-     *
-     * @param {String} [text] The text to be displayed on the screen if no imageUrl is specified.
-     * @param {String} [imageUrl] The source location for an image
-     * @param {String} [link] A URL location for which the credit will be hyperlinked
-     * @param {String} [showInPopup=true] If false, the credit will be visible in the main credit container.  Otherwise, it will appear in a popover
+     * @param {Object} [options] An object with the following properties
+     * @param {String} [options.text] The text to be displayed on the screen if no imageUrl is specified.
+     * @param {String} [options.imageUrl] The source location for an image
+     * @param {String} [options.link] A URL location for which the credit will be hyperlinked
+     * @param {String} [options.showOnScreen=false] If true, the credit will be visible in the main credit container.  Otherwise, it will appear in a popover
      *
      * @alias Credit
      * @constructor
@@ -28,7 +30,20 @@ define([
      * //Create a credit with a tooltip, image and link
      * var credit = new Cesium.Credit('Cesium', '/images/cesium_logo.png', 'http://cesiumjs.org/');
      */
-    function Credit(text, imageUrl, link, showInPopup) {
+    function Credit(options, imageUrl, link) {
+        var text;
+        var showOnScreen;
+        if (typeof options !== 'object') {
+            deprecationWarning('Credit parameters', 'The Credit text, imageUrl and link parameters have been replaced by a single options object parameter with text, imageUrl and link properties. Use of the old parameters will be removed in Cesium 1.41');
+            text = options;
+            showOnScreen = false;
+        } else {
+            text = options.text;
+            imageUrl = options.imageUrl;
+            link = options.link;
+            showOnScreen = defaultValue(options.showOnScreen, false);
+        }
+
         var hasLink = (defined(link));
         var hasImage = (defined(imageUrl));
         var hasText = (defined(text));
@@ -48,7 +63,7 @@ define([
         this._link = link;
         this._hasLink = hasLink;
         this._hasImage = hasImage;
-        this._showInPopup = defaultValue(showInPopup, true);
+        this._showOnScreen = showOnScreen;
 
         // Credits are immutable so generate an id to use to optimize equal()
         var id;
@@ -120,9 +135,9 @@ define([
          * @type {Boolean}
          * @readonly
          */
-        showInPopup : {
+        showOnScreen : {
             get : function() {
-                return this._showInPopup;
+                return this._showOnScreen;
             }
         }
     });